fix(tests): remove error handler from middlware
