+++ src/core/middleware.py
@@ -10,7 +10,7 @@ This module provides middleware components for:
 import time
 from typing import List, Optional
 
-from flask import Flask, g, jsonify, request
+from flask import Flask, g, request
 from flask_jwt_extended import decode_token, get_jwt_identity
 
 from .structured_logging import (
@@ -253,125 +253,3 @@ class RequestContextMiddleware:
                 request_id=g.get("request_id", "unknown"),
                 duration_ms=round((time.time() - g.request_metadata["timestamp"]) * 1000, 2),
             )
-
-
-def create_error_handler(app: Flask) -> None:
-    """Create global error handlers with structured logging.
-
-    Args:
-        app: Flask application instance
-    """
-    logger = get_logger(__name__)
-
-    @app.errorhandler(400)
-    def bad_request(error):
-        """Handle 400 Bad Request errors."""
-        logger.warning(
-            "Bad request error",
-            error=str(error),
-            status_code=400,
-            path=request.path,
-            method=request.method,
-        )
-        return (
-            jsonify(
-                {
-                    "error": "Bad Request",
-                    "message": str(error),
-                    "status_code": 400,
-                    "correlation_id": g.get("correlation_id", "unknown"),
-                }
-            ),
-            400,
-        )
-
-    @app.errorhandler(401)
-    def unauthorized(error):
-        """Handle 401 Unauthorized errors."""
-        logger.warning(
-            "Unauthorized access attempt",
-            error=str(error),
-            status_code=401,
-            path=request.path,
-            method=request.method,
-            client_ip=request.environ.get("REMOTE_ADDR", "unknown"),
-        )
-        return (
-            jsonify(
-                {
-                    "error": "Unauthorized",
-                    "message": "Authentication required",
-                    "status_code": 401,
-                    "correlation_id": g.get("correlation_id", "unknown"),
-                }
-            ),
-            401,
-        )
-
-    @app.errorhandler(403)
-    def forbidden(error):
-        """Handle 403 Forbidden errors."""
-        logger.warning(
-            "Forbidden access attempt",
-            error=str(error),
-            status_code=403,
-            path=request.path,
-            method=request.method,
-            user_id=g.get("user_id", "anonymous"),
-        )
-        return (
-            jsonify(
-                {
-                    "error": "Forbidden",
-                    "message": "Access denied",
-                    "status_code": 403,
-                    "correlation_id": g.get("correlation_id", "unknown"),
-                }
-            ),
-            403,
-        )
-
-    @app.errorhandler(404)
-    def not_found(error):
-        """Handle 404 Not Found errors."""
-        logger.info(
-            "Resource not found",
-            error=str(error),
-            status_code=404,
-            path=request.path,
-            method=request.method,
-        )
-        return (
-            jsonify(
-                {
-                    "error": "Not Found",
-                    "message": "The requested resource was not found",
-                    "status_code": 404,
-                    "correlation_id": g.get("correlation_id", "unknown"),
-                }
-            ),
-            404,
-        )
-
-    @app.errorhandler(500)
-    def internal_error(error):
-        """Handle 500 Internal Server errors."""
-        logger.error(
-            "Internal server error",
-            error=str(error),
-            status_code=500,
-            path=request.path,
-            method=request.method,
-            exc_info=True,
-        )
-        return (
-            jsonify(
-                {
-                    "error": "Internal Server Error",
-                    "message": "An unexpected error occurred",
-                    "status_code": 500,
-                    "correlation_id": g.get("correlation_id", "unknown"),
-                }
-            ),
-            500,
-        )
